listener

+++ examples/interactive_segmentation/android/shared/src/main/java/com/google/mediapipe/tasks/vision/provider/DownloadStatus.kt
@@ -1,7 +1,7 @@
 package com.google.mediapipe.tasks.vision.provider
 sealed class DownloadStatus {
     object Idle : DownloadStatus()
-    data class Downloading(val progress: Int) : DownloadStatus()
+    data class Downloading(val progress: Float) : DownloadStatus()
     data class Completed(val path: String) : DownloadStatus()
     data class Failed(val errorCode: Int) : DownloadStatus()
 }
